+++ apps/expiration/drizzle/0000_expiration_initial.sql
@@ -2,7 +2,7 @@ CREATE SCHEMA IF NOT EXISTS "expiration";
 --> statement-breakpoint
 CREATE TABLE "expiration"."inbox" (
 	"id" uuid PRIMARY KEY DEFAULT gen_random_uuid() NOT NULL,
-	"event_id" text NOT NULL,
+	"event_id" uuid NOT NULL,
 	"subject" text NOT NULL,
 	"consumed_at" timestamp with time zone DEFAULT now() NOT NULL,
 	CONSTRAINT "inbox_event_subject_uq" UNIQUE("event_id","subject")
