+++ apps/expiration/drizzle/0001_expiration_inbox_event_id_text.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "expiration"."inbox" ALTER COLUMN "event_id" SET DATA TYPE text USING "event_id"::text;
