+++ apps/expiration/drizzle/meta/0001_snapshot.json
@@ -1,6 +1,6 @@
 {
-  "id": "ee22d1e7-ac41-411f-b721-a8864f4908ef",
-  "prevId": "00000000-0000-0000-0000-000000000000",
+  "id": "a1111111-1111-4111-8111-111111111111",
+  "prevId": "ee22d1e7-ac41-411f-b721-a8864f4908ef",
   "version": "7",
   "dialect": "postgresql",
   "tables": {
@@ -17,7 +17,7 @@
         },
         "event_id": {
           "name": "event_id",
-          "type": "uuid",
+          "type": "text",
           "primaryKey": false,
           "notNull": true
         },
