+++ apps/orders/drizzle/0002_orders_inbox.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "orders"."inbox" (
 	"id" uuid PRIMARY KEY DEFAULT gen_random_uuid() NOT NULL,
-	"event_id" text NOT NULL,
+	"event_id" uuid NOT NULL,
 	"subject" text NOT NULL,
 	"consumed_at" timestamp with time zone DEFAULT now() NOT NULL,
 	CONSTRAINT "inbox_event_subject_uq" UNIQUE("event_id","subject")
