+++ apps/orders/drizzle/0003_orders_inbox_event_id_text.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "orders"."inbox" ALTER COLUMN "event_id" SET DATA TYPE text USING "event_id"::text;
