+++ apps/orders/drizzle/meta/0002_snapshot.json
@@ -81,7 +81,7 @@
         },
         "event_id": {
           "name": "event_id",
-          "type": "text",
+          "type": "uuid",
           "primaryKey": false,
           "notNull": true
         },
