+++ apps/orders/drizzle/meta/0003_snapshot.json
@@ -1,12 +1,12 @@
 {
-  "id": "331a1bd3-a125-4b0e-a2b4-6e724bf243d9",
-  "prevId": "05569616-f136-428a-8a50-aa2a080779b9",
+  "id": "a2222222-2222-4222-8222-222222222222",
+  "prevId": "7a8f3258-c5e2-40ad-9de8-a29c52aaa051",
   "version": "7",
   "dialect": "postgresql",
   "tables": {
-    "tickets.tickets": {
-      "name": "tickets",
-      "schema": "tickets",
+    "orders.orders": {
+      "name": "orders",
+      "schema": "orders",
       "columns": {
         "id": {
           "name": "id",
@@ -15,33 +15,33 @@
           "notNull": true,
           "default": "gen_random_uuid()"
         },
-        "seller_id": {
-          "name": "seller_id",
+        "buyer_id": {
+          "name": "buyer_id",
           "type": "text",
           "primaryKey": false,
           "notNull": true
         },
-        "title": {
-          "name": "title",
+        "ticket_id": {
+          "name": "ticket_id",
+          "type": "uuid",
+          "primaryKey": false,
+          "notNull": true
+        },
+        "quantity": {
+          "name": "quantity",
+          "type": "integer",
+          "primaryKey": false,
+          "notNull": true
+        },
+        "status": {
+          "name": "status",
           "type": "text",
           "primaryKey": false,
           "notNull": true
         },
-        "quantity_total": {
-          "name": "quantity_total",
-          "type": "integer",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "quantity_available": {
-          "name": "quantity_available",
-          "type": "integer",
-          "primaryKey": false,
-          "notNull": true
-        },
-        "unit_price_cents": {
-          "name": "unit_price_cents",
-          "type": "integer",
+        "expires_at": {
+          "name": "expires_at",
+          "type": "timestamp with time zone",
           "primaryKey": false,
           "notNull": true
         },
@@ -68,9 +68,9 @@
       "checkConstraints": {},
       "isRLSEnabled": false
     },
-    "tickets.inbox": {
+    "orders.inbox": {
       "name": "inbox",
-      "schema": "tickets",
+      "schema": "orders",
       "columns": {
         "id": {
           "name": "id",
@@ -81,7 +81,7 @@
         },
         "event_id": {
           "name": "event_id",
-          "type": "uuid",
+          "type": "text",
           "primaryKey": false,
           "notNull": true
         },
@@ -113,9 +113,9 @@
       "checkConstraints": {},
       "isRLSEnabled": false
     },
-    "tickets.outbox": {
+    "orders.outbox": {
       "name": "outbox",
-      "schema": "tickets",
+      "schema": "orders",
       "columns": {
         "id": {
           "name": "id",
@@ -173,7 +173,7 @@
   },
   "enums": {},
   "schemas": {
-    "tickets": "tickets"
+    "orders": "orders"
   },
   "sequences": {},
   "roles": {},
