+++ apps/orders/drizzle/meta/_journal.json
@@ -22,6 +22,13 @@
       "when": 1779413649145,
       "tag": "0002_orders_inbox",
       "breakpoints": true
+    },
+    {
+      "idx": 3,
+      "version": "7",
+      "when": 1779413649146,
+      "tag": "0003_orders_inbox_event_id_text",
+      "breakpoints": true
     }
   ]
 }
