+++ packages/contracts/src/subjects.ts
@@ -3,4 +3,5 @@ export const ORDER_CREATED_V1 = "order.created.v1";
 export const ORDER_EXPIRED_V1 = "order.expired.v1";
 export const ORDER_RESERVATION_RELEASED_V1 = "order.reservation_released.v1";
 
+export const TICKETS_STREAM = "TICKETS";
 export const ORDERS_STREAM = "ORDERS";
