+++ scripts/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "@tix/config/tsconfig.base.json",
   "include": ["*.ts"],
   "compilerOptions": {
+    "allowImportingTsExtensions": true,
     "noEmit": true
   }
 }
